% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.weighted.yield.R
\name{area.weighted.yield}
\alias{area.weighted.yield}
\title{area.weighted.yield}
\usage{
area.weighted.yield(data, nbands=32, year=2000, landusefile,
gridfile)
}
\arguments{
\item{data}{lpj object}

\item{nbands}{integer. Number of bands in yield and landuse input file from
LPJmL run.}

\item{year}{integer. Landuse input file will be read only for this year and
used as weight.}

\item{landusefile}{character. Name of landuse input file used in LPJmL run
incl. name of folder}

\item{gridfile}{character. Name of grid output file from LPJmL run incl.
name of folder}
}
\value{
The function returns a list, with 2 lpj objects: "data" containing
the yield and "croparea" containing the crop area in ha
}
\description{
Function calculates cell-and crop-specific area-weighted mean yield from
rainfed and irrigated yield from a lpj object.
}
\details{
Default for nbands is 32 for LPJmL runs with 16 crops (incl. sugarcane and
bioenergy), default for year is 2000
}
\examples{
 \dontrun{mean.yield<-area.weighted.yield(yield,landusefile="D:/LPJ/input/cft1700_2005_16cfts_SR.bin",
					     gridfile="D:/LPJ/run1/grid.bin")}
 
}
\seealso{
\code{\link{readLPJ}}, \code{\link{new.lpj}}
}
\author{
Katharina Waha
}
