% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpj-class.R
\docType{class}
\name{lpj-class}
\alias{lpj-class}
\alias{as.lpj}
\alias{is.lpj}
\alias{[,lpj-method}
\alias{[,lpj,ANY,ANY-method}
\alias{[<-,lpj-method}
\alias{[<-,lpj,ANY,ANY-method}
\title{Class "lpj" ~~~}
\arguments{
\item{x}{An object that should be either tested or transformed as/to an
LPJ-object.}
}
\description{
The LPJ class is a data format for cellular LPJ data with a close
relationship to the array data format. \code{is.lpj} tests if \code{x} is an
LPJ-object, \code{as.lpj} transforms \code{x} to an LPJ-object (if
possible).
}
\examples{

showClass("lpj")

}
\seealso{
\code{\link{readLPJ}}, \code{\link{read.LPJ_input}}
}
\author{
Jan Philipp Dietrich, Susanne Rolinski, Katharina Waha, Benjamin
Bodirsky
}
\keyword{classes}
