% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjclass-package.R
\docType{package}
\name{lpjclass-package}
\alias{lpjclass-package}
\alias{lpjclass}
\title{LPJ Class Functions}
\description{
Package containing the LPJ-Object-Class together with relevant functions and
methods.
}
\details{
\tabular{ll}{ Package: \tab lpjclass\cr Type: \tab Package\cr Version: \tab
1.11\cr Date: \tab 2016-09-20\cr License: \tab LGPL-3\cr LazyLoad: \tab
yes\cr }
}
\author{
Susanne Roliniski, Benjamin Bodirsky, Katharina Waha, Jan Philipp
Dietrich

Maintainer: Susanne Rolinski <rolinski@pik-potsdam.de>
}
