% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjclassdata.R
\name{lpjclassdata}
\alias{lpjclassdata}
\title{lpjclassdata}
\description{
General LPJ-class-dataset
}
\details{
Please do not directly access that data. It should be only used by library
functions.
}
\author{
Jan Philipp Dietrich
}
