% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.lpj.R
\name{new.lpj}
\alias{new.lpj}
\title{new.lpj}
\usage{
new.lpj(cells_and_regions,years,names,irrigation)
}
\arguments{
\item{cells_and_regions}{dimnames for cells in the format
"country.cellnumber"}

\item{years}{dimnames for years in the format "yXXXX"}

\item{names}{dimnames for names}

\item{irrigation}{dimnames for irrigation}
}
\value{
an empty lpj object filled with NAs, with the given dimnames
}
\description{
Creates a new LPJ object
}
\examples{

 cells <- c("Russia.1","Russia.2")
 years <- "y1995"
 categories <- c("TeCe","TrRi")
 irrigation <- c("rainfed","irrigated")
 a <- new.lpj(cells_and_regions=cells,years=years,names=categories,irrigation=irrigation)
 a

}
\seealso{
\code{\linkS4class{lpj}},\code{\link{readLPJ}}
}
\author{
Benjamin Bodirsky, Katharina Waha
}
