% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.LPJ_input.R
\name{read.LPJ_input}
\alias{read.LPJ_input}
\title{read.LPJ_input}
\usage{
read.LPJ_input(
  file_name,
  out_years = c("y1995", "y2005"),
  namesum = FALSE,
  four_d = FALSE,
  ncells = 59199,
  swap = "little",
  rule4binary = NULL
)
}
\arguments{
\item{file_name}{Filename with extension and folder}

\item{out_years}{years to be red out in the form of a vector of year
strings, e.g. c(y1995,y2005)}

\item{namesum}{if true, all bands/colums of the dataset are added. Sometimes
useful to handle large datasets. Overwrites four_d}

\item{four_d}{if true, it is assumed that data exists for both rainfed and
irrigated crops.}

\item{ncells}{number of cells}

\item{swap}{Depends on the binary format of the data}

\item{rule4binary}{Rule based transformation of the outputs into binary values}
}
\value{
\item{x}{LPJ-object}
}
\description{
Reads a LPJmL input file and converts it to a LPJ-object
}
\details{
This function reads in LPJ-input files, using its header information. So
far, tested for landuse input. `rule4binary` combined with `namesum = T` allow for the rule based counting of events.
}
\seealso{
\code{\link{readLPJ}}
}
\author{
Benjamin Bodirsky, Susanne Roliniski, Marcos Alves
}
