% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round-methods.R
\docType{methods}
\name{round-methods}
\alias{round-methods}
\alias{round,lpj-method}
\title{Round-method for LPJ objects}
\description{
Round-method for LPJ-objects respectively. Works exactly as for arrays.
}
\section{Methods}{
 \describe{

\item{x = "lpj"}{ works as round(x) for arrays. }

}
}

