% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatis.R
\name{whatis}
\alias{whatis}
\title{whatis}
\usage{
whatis(x)
}
\arguments{
\item{x}{a character with quotes.}
}
\value{
a list with four list elements on: LPJmLName - the LPJmL name,
CropAreaLPJmL - the crops whose cultivated areas are contained in the LPJmL
landuse input file, CropGroupMAgPIE - the FAO crops whose areas and yields
are compared to MAgPIE areas and yield, RepresentativeCropLPJmL - the
parameters from this crop are used for the whole group in LPJmL.
}
\description{
Helps to understand cropnames and abbreviation used in LPJmL and MAgPIE
}
\examples{

 whatis("tece")

}
\seealso{
\code{\link{readLPJ}},\code{\link{read.LPJ_input}}
}
\author{
Katharina Waha
}
